requirements.txt
Adafruit-PlatformDetect==3.5.0
adafruit-python-shell==1.3.1
args==0.1.0
clint==0.5.1
RPi.bme280==0.2.3
RPi.GPIO==0.7.0
smbus2==0.4.1
adafruit-blinka==6.12.0
